refactor: replace alertRequestData with ModalRequestData component
Alert 시스템에서 독립적인 모달 컴포넌트로 요청 데이터 뷰어 분리.
module context를 활용한 openRequestData() 함수로 변경.
